requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
supabase==2.12.0
requests==2.32.3
gunicorn==21.2.0
Werkzeug==3.0.3
openpyxl==3.1.5